requirements.txt
Django==3.1.4
pytesseract==0.3.7
Pillow==8.0.1
opencv-python==4.4.0.46
ftfy==5.8